fix(native): mac client icon refresh
